Chore: Inline TextareaInput into Textarea
